pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wcc26-feistel-structures"
version = "0.1.0"
description = "Experimental code accompanying the paper on 3-round Feistel Sboxes (WCC 2026)"
readme = "README.md"
requires-python = ">=3.8"

authors = [
    { name = "Merlin Fruchon", email = "merlin.fruchon@inria.fr" }
]

dependencies = [
    "sboxU @ git+https://github.com/lpp-crypto/sboxU.git@v2"
]

[tool.setuptools]
packages = []
